+++ assets/js/theme/category.js
@@ -5,6 +5,13 @@ import FacetedSearch from './common/faceted-search';
 
 export default class Category extends CatalogPage {
     onReady() {
+        $('[data-button-type="add-cart"]').on('click', (e) => {
+            $(e.currentTarget).next().attr({
+                role: 'status',
+                'aria-live': 'polite',
+            });
+        });
+
         compareProducts(this.context.urls);
 
         if ($('#facetedSearch').length > 0) {